Link author name in catch-up recap (parity with daily summary)
The catch-up recap showed the question author as plain text and omitted
the Editorial badge for house questions, unlike the daily-five summary
which links the human author's name to their profile. The cause was that
authorId was never threaded through the catch-up pipeline.

Carry authorId from the catch-up query (CatchupQueueItem) through the GET
route and the client record (CatchupBatchRecord), following the same path
authorName already takes, then render it via a shared AuthorName component
extracted from the summary page, plus EditorialBadge for house questions.
